Extract shared utilities into @gseurat/vfx-utils package
Move duplicated code from Méliès and Bricklayer into a new shared
package:

- catmullRom.ts: Catmull-Rom spline evaluation (was identical in both)
- plyLoader.ts: Binary PLY parser with SH DC + RGB support (from Méliès)
- wasmLoader.ts: Shared WASM simulation module loader (was duplicated
  3x with slight variations)

Both apps now import from @gseurat/vfx-utils. The inline PLY parser
in Bricklayer's VfxRenderer (which only handled float properties) is
replaced by the more complete shared version in a follow-up.
