+++ tools/apps/melies/package.json
@@ -11,6 +11,7 @@
   "dependencies": {
     "@gseurat/engine-client": "workspace:*",
     "@gseurat/simulation-wasm": "workspace:*",
+    "@gseurat/vfx-utils": "workspace:*",
     "@gseurat/test-harness": "workspace:*",
     "@gseurat/ui-kit": "workspace:*",
     "@react-three/drei": "^9.117.0",
