+++ tools/packages/vfx-utils/src/catmullRom.ts
