+++ tools/packages/vfx-utils/src/index.ts
@@ -1,0 +1,3 @@
+export * from './catmullRom.js';
+export * from './plyLoader.js';
+export * from './wasmLoader.js';
